feat: implement dynamic chat session routing with /chat/{session-id}
- Add dynamic routing for individual chat sessions on /chat/{session-id} URLs
- Create new [sessionId].vue page for handling session-specific routes
- Update ChatSessionList to navigate using new URL structure
- Modify session creation and selection to use router navigation
- Add proper error handling for model loading during navigation
- Fix AbortError handling in useModels for graceful request cancellation
- Improve concurrent model loading with better timeout and caching logic
- Remove playground route reference from navigation menu
- Add SSR compatibility with ClientOnly wrapper for Chat component
