+++ composables/useMenus.ts
@@ -11,7 +11,6 @@ export function useMenus() {
         { label: t('menu.knowledgeBases'), icon: 'i-heroicons-book-open', to: '/knowledgebases' },
         { label: t('menu.chat'), icon: 'i-iconoir-chat-lines', to: '/chat' },
         { label: t('menu.realtime'), icon: 'i-iconoir-microphone', to: '/realtime' },
-        { label: t('menu.playground'), icon: 'i-heroicons-beaker', to: '/playground' },
         { label: t('menu.settings'), icon: 'i-heroicons-cog-6-tooth', to: '/settings' }
     ])
 }
